refactoring inferencee cliping and earlystopping
